+++ libs/angular-vanilla/src/lib/utils/templates/errors.component.ts
@@ -7,7 +7,7 @@ import * as Core from '@golemui/core';
   selector: '[gui-errors]',
   imports: [CommonModule],
   template: `@for (error of templateData().errors; track $index) {
-    <li class="gui-validator__error" role="status" [attr.data-cy]="uid() + '_validator-error'">
+    <li class="gui-validator__error" role="alert" [attr.data-cy]="uid() + '_validator-error'">
       {{ error }}
     </li>
   }`,