+++ libs/react/src/lib/hooks/useControlField.ts
@@ -1,5 +1,6 @@
 import * as Core from '@golemui/core';
 import { useCallback, useEffect, useState } from 'react';
+import { combineLatest } from 'rxjs';
 import { useReactFormContext } from '../ReactFormContext';
 import { useTemplateData, WithFlattenedProps } from './internal/useExtraProps';
 
@@ -53,11 +54,16 @@ export function useControlField<T, ExtraProps extends Record<string, any>>(
 
   // Listen to the validation stream for this control
   useEffect(() => {
-    const sub = formContext.store.state$
-      .pipe(Core.validationByPath$(field.path))
-      .subscribe((validation) => {
-        setErrors(validation?.status?.errors || []);
-      });
+    const validation$ = formContext.store.state$.pipe(Core.validationByPath$(field.path));
+    const injectedValidation$ = formContext.store.state$.pipe(
+      Core.injectedValidationByPath$(field.path),
+    );
+
+    const sub = combineLatest([validation$, injectedValidation$]).subscribe(
+      ([validation, injectedValidation]) => {
+        setErrors([...(validation?.status?.issues ?? []), ...(injectedValidation ?? [])]);
+      },
+    );
     return () => sub.unsubscribe();
   }, [field, formContext.store]);
 
@@ -95,6 +101,16 @@ export function useControlField<T, ExtraProps extends Record<string, any>>(
     [field, formContext],
   );
 
+  const injectValidationIssues = useCallback(
+    (issues: string[] | null) => {
+      formContext.store.dispatch({
+        type: 'INJECT_VALIDATION_ISSUES',
+        payload: { path: field.path, issues },
+      });
+    },
+    [field, formContext],
+  );
+
   const onBlur = useCallback(() => {
     formContext.store.dispatch({
       type: 'ATTEMPT_VALIDATION',
@@ -110,5 +126,6 @@ export function useControlField<T, ExtraProps extends Record<string, any>>(
     isTouched,
     onValueChanged,
     onBlur,
+    injectValidationIssues,
   };
 }