+++ libs/lit-vanilla/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.15.2 (2026-01-03)
+
+### 🩹 Fixes
+
+- **cypress:** rename formError to formHealth ([0aee3df](https://github.com/golemui/formforge/commit/0aee3df))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
 ## 0.15.1 (2025-12-27)
 
 ### 🩹 Fixes