+++ libs/react-vanilla/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.15.2 (2026-01-03)
+
+### 🩹 Fixes
+
+- **cypress:** rename formError to formHealth ([0aee3df](https://github.com/golemui/formforge/commit/0aee3df))
+- **react:** rename formError to formHealth ([3eba8ac](https://github.com/golemui/formforge/commit/3eba8ac))
+
+### ❤️ Thank You
+
+- Mud Scientist
+
 ## 0.15.1 (2025-12-27)
 
 ### 🩹 Fixes